Added entity rendering
Entity rendering implemented, made changes to 'MainStage' and finished
testing 'TriggerFactory'.
Testing entity rendering and states next.

+++ GDX/core/src/com/silvercarnival/mapobjects/TriggerFactory.java
@@ -79,7 +79,7 @@ public class TriggerFactory implements ObjectFactory {
 			}
 		}
 		
-		entity.add(new PositionComponent(new Vector2(e.getInt("x"),e.getInt("y"))));
+		entity.add(new PositionComponent(new Vector2(e.getFloat("x"),e.getFloat("y"))));
 		entity.add(new SensorComponent(e.getInt("width"), e.getFloat("height"), world));
 		entity.add(new TriggerComponent(type, delay, data));
 		entity.add(new DebugRenderTag());//for debugging
